backend/requirements.txt
fastapi
uvicorn
authx
httpx
asyncio
supabase
python-dotenv
py-fast-grpc